python/requirements.txt
minizinc
pyswip
forsyde-io-python
numpy
